// scrapper/app/enums/LegislationUrlPrefixes.ts
export enum LegislationUrlPrefixes {
    Laws = "#kanunlar",
    DecreesWithForceOfLaw = "#kanunHukmundeKararnameler",
    PresidentialDecrees = "#cumhurbaskanligiKararnameleri",
    Regulations = "#tuzukler",
    PresidentialDecisions = "#cumhurbaskaniKararlari",
    PresidentialCirculars = "#cumhurbaskaniGenelgeleri",
    AdministrativeRegulations = "#yonetmelikler",
    Notifications = "#tebligler"
}