// scrapper/app/enums/LegislationForms.ts
export enum LegislationForms {
    Laws = "#kanunlar_form",
    DecreesWithForceOfLaw = "#kanunHukmundeKararnameler_form",
    PresidentialDecrees = "#cumhurbaskanligiKararnameleri_form",
    Regulations = "#tuzukler_form",
    PresidentialDecisions = "#cumhurbaskaniKararlari_form",
    PresidentialCirculars = "#cumhurbaskaniGenelgeleri_form",
    AdministrativeRegulations = "#yonetmelikler_form",
    Notifications = "#tebligler_form"
}